added ability to specify number of cores for multiprocessing
If to_use_parallelization is True, the user now has the option to specify num_cores. By default, num_cores is None, which defaults to the currently expected behaviour of using multiprocessing.cpu_count().

This change is motivated by the fact that on HPC clusters, cpu_count() may return more cores than the user is allocated.